git commit -am 'Add some snippet'
This is the first time I use git. I might be wrong.

But I edited something in the snippet.json. The controller. And I added //$ so the $ can show up in the result. Check it out.

I will edit more if I did nothing wrong in the json file, or in this git activities.